pyproject.toml
[project]
name = "llm-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "gradio>=5.46.1",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.12",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.8",
    "langchain-openai>=0.3.33",
    "langchain-tavily>=0.2.11",
    "langchain-text-splitters>=0.3.11",
    "langfuse>=3.5.0",
    "langgraph-cli[inmem]>=0.4.2",
    "langgraph[mermaid]>=0.6.7",
    "langsmith>=0.4.29",
    "lark>=1.3.0",
    "playwright>=1.55.0",
    "pypdf>=6.1.0",
    "python-dotenv>=1.1.1",
    "ragas==0.3.1",
    "ranx>=0.3.21",
    "rapidfuzz>=3.14.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.0",
    "plotly>=5.17.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
]
